# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_patrick_cocobase_internal_models.CollectionCreateRequest:
    properties:
      name:
        type: string
      permissions:
        $ref: '#/definitions/github_com_patrick_cocobase_internal_models.CollectionPermissionsRequest'
      webhook_url:
        type: string
    required:
    - name
    type: object
  github_com_patrick_cocobase_internal_models.CollectionPermissionsRequest:
    properties:
      create:
        items:
          type: string
        type: array
      delete:
        items:
          type: string
        type: array
      read:
        items:
          type: string
        type: array
      update:
        items:
          type: string
        type: array
    type: object
  github_com_patrick_cocobase_internal_models.CollectionResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        $ref: '#/definitions/github_com_patrick_cocobase_internal_models.Permissions'
      project_id:
        type: string
      webhook_url:
        type: string
    type: object
  github_com_patrick_cocobase_internal_models.CollectionUpdateRequest:
    properties:
      name:
        type: string
    type: object
...
  github_com_patrick_cocobase_internal_models.DocumentResponse:
    properties:
      collection_id:
        type: string
      created_at:
        type: string
      data:
        additionalProperties: true
        type: object
      id:
        type: string
    type: object
...
  github_com_patrick_cocobase_internal_models.Permissions:
    properties:
      create:
        items:
          type: string
        type: array
      delete:
        items:
          type: string
        type: array
      read:
        items:
          type: string
        type: array
      update:
        items:
          type: string
        type: array
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@cocobase.io
    name: Cocobase Support
  description: Backend as a Service with flexible collections and document management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cocobase API
  version: "1.0"
paths:
...
  /collections:
    post:
      consumes:
      - application/json
      description: Create a new collection with optional permissions
      parameters:
      - description: Collection data
        in: body
        name: collection
        required: true
        schema:
          $ref: '#/definitions/github_com_patrick_cocobase_internal_models.CollectionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_patrick_cocobase_internal_models.CollectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new collection
      tags:
      - Collections
  /collections/{id}:
    delete:
      description: Delete a collection and all its documents
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a collection
      tags:
      - Collections
    get:
      description: Retrieve a collection by its ID or name
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_patrick_cocobase_internal_models.CollectionResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get a collection
      tags:
      - Collections
    patch:
      consumes:
      - application/json
      description: Update a collection. Use helper function, invalidate cache after
        update.
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: collection
        required: true
        schema:
          $ref: '#/definitions/github_com_patrick_cocobase_internal_models.CollectionUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_patrick_cocobase_internal_models.CollectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update collection
      tags:
      - Collections
...
  /collections/{id}/documents/upload:
    post:
      consumes:
      - multipart/form-data
      description: Create a document and automatically upload files to S3, adding
        URLs to the document
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      - description: Document data as JSON string
        in: formData
        name: data
        required: true
        type: string
      - description: Field name to store file URL (defaults to 'file_url')
        in: formData
        name: file_field
        type: string
      - description: File to upload
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_patrick_cocobase_internal_models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "413":
          description: Request Entity Too Large
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create document with file upload
      tags:
      - Documents
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and API key.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"